Создан обработчик dispatch, переименованы страницы, убраны выполненные задачи.
Создан обработчик dispatch, переименованы страницы, убраны выполненные задачи.
