+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.11.8",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.6.0",
+    "axios": "^0.21.1",
     "node-sass": "^4.14.1",
     "normalize.css": "^8.0.1",
     "react": "^17.0.1",
