+++ src/index.js
@@ -5,7 +5,7 @@ import reportWebVitals from './reportWebVitals';
 import {Provider} from "react-redux";
 
 // Components
-import Navigation from './pages/Navigation';
+import NavigationPage from './pages/NavigationPage';
 import store from './store/store'
 
 // Style
@@ -21,7 +21,7 @@ import 'normalize.css';
 render((
     <Provider store={store}>
         <BrowserRouter>
-            <Navigation/>
+            <NavigationPage/>
         </BrowserRouter>
     </Provider>
 ), document.getElementById('root'));
